+++ apps/web/ui/analytics/toggle.tsx
@@ -5,16 +5,11 @@ import {
   BadgeTooltip,
   BlurImage,
   ExpandingArrow,
-  IconMenu,
-  Popover,
   Tick,
-  Tooltip,
-  TooltipContent,
   useRouterStuff,
   useScroll,
 } from "@dub/ui";
 import {
-  APP_DOMAIN,
   DUB_LOGO,
   GOOGLE_FAVICON_URL,
   SHORT_DOMAIN,
@@ -22,7 +17,6 @@ import {
   getApexDomain,
   linkConstructor,
 } from "@dub/utils";
-import { Calendar, ChevronDown, Lock } from "lucide-react";
 import Link from "next/link";
 import { useSearchParams } from "next/navigation";
 import punycode from "punycode/";
@@ -31,21 +25,13 @@ import { AnalyticsContext } from ".";
 import ExportButton from "./export-button";
 import SharePopover from "./share-popover";
 import TagSelector from "./tag-selector";
+import DateRangePicker from "./date-range-picker";
 
 export default function Toggle() {
-  const { queryParams } = useRouterStuff();
-
-  const { basePath, domain, key, url, interval, admin } =
-    useContext(AnalyticsContext);
-
-  const [openDatePopover, setOpenDatePopover] = useState(false);
-
-  const selectedInterval = useMemo(() => {
-    return INTERVALS.find((s) => s.value === interval) || INTERVALS[1];
-  }, [interval]);
+  const { basePath, domain, key, url, admin } = useContext(AnalyticsContext);
 
   const scrolled = useScroll(80);
-  const { name, plan, logo } = useWorkspace();
+  const { name, logo } = useWorkspace();
   const { allActiveDomains, primaryDomain } = useDomains();
 
   const isPublicStatsPage = basePath.startsWith("/stats");
@@ -104,80 +90,7 @@ export default function Toggle() {
         <div className="flex w-full items-center justify-end gap-2">
           {!isPublicStatsPage && key && <SharePopover />}
           {!isPublicStatsPage && !key && <TagSelector />}
-          <Popover
-            content={
-              <div className="grid w-full p-2 md:w-48">
-                {INTERVALS.map(({ display, value }) =>
-                  (value === "all" || value === "90d") &&
-                  (!plan || plan === "free") &&
-                  !admin ? (
-                    <Tooltip
-                      key={value}
-                      content={
-                        <TooltipContent
-                          title={`${display} stats can only be viewed on a Pro plan or higher. Upgrade now to view all-time stats.`}
-                          cta="Upgrade to Pro"
-                          {...(isPublicStatsPage
-                            ? {
-                                href: APP_DOMAIN,
-                              }
-                            : {
-                                onClick: () => {
-                                  setOpenDatePopover(false);
-                                  queryParams({
-                                    set: {
-                                      upgrade: "pro",
-                                    },
-                                  });
-                                },
-                              })}
-                        />
-                      }
-                    >
-                      <div className="flex w-full cursor-not-allowed items-center justify-between space-x-2 rounded-md p-2 text-sm text-gray-400">
-                        <p>{display}</p>
-                        <Lock className="h-4 w-4" aria-hidden="true" />
-                      </div>
-                    </Tooltip>
-                  ) : (
-                    <Link
-                      key={value}
-                      href={
-                        queryParams({
-                          set: { interval: value },
-                          getNewPath: true,
-                        }) as string
-                      }
-                      className="flex w-full items-center justify-between space-x-2 rounded-md p-2 hover:bg-gray-100 active:bg-gray-200"
-                    >
-                      <p className="text-sm">{display}</p>
-                      {selectedInterval.value === value && (
-                        <Tick className="h-4 w-4" aria-hidden="true" />
-                      )}
-                    </Link>
-                  ),
-                )}
-              </div>
-            }
-            openPopover={openDatePopover}
-            setOpenPopover={setOpenDatePopover}
-          >
-            <button
-              onClick={() => setOpenDatePopover(!openDatePopover)}
-              className="flex w-full items-center justify-between space-x-2 truncate rounded-md bg-white px-3 py-2.5 shadow transition-all hover:shadow-md md:w-48"
-            >
-              <IconMenu
-                text={selectedInterval.display}
-                icon={<Calendar className="h-4 w-4 flex-shrink-0" />}
-              />
-              <ChevronDown
-                className={`h-4 w-4 text-gray-400 ${
-                  openDatePopover ? "rotate-180 transform" : ""
-                } transition-all duration-75`}
-              />
-            </button>
-          </Popover>
-
+          <DateRangePicker />
           {!isPublicStatsPage && <ExportButton />}
         </div>
       </div>
